refactor(path): make Path depend directly on Adapter concrete class
- remove Navigator interface and rename Adapter to Navigator
